# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phantom Mask API",
        "description": "API for managing pharmacies, masks, and purchases",
        "contact": {
            "email": "rouyucc@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/purchases": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "處理使用者購買口罩的交易",
                "description": "處理顧客從藥局購買口罩的交易。驗證資金是否充足，更新顧客餘額和藥局餘額，並建立購買記錄。",
                "operationId": "buyMask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "pharmacy_id",
                                    "mask_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "進行購買的顧客 ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pharmacy_id": {
                                        "description": "銷售口罩的藥局 ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mask_id": {
                                        "description": "被購買的口罩 ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "購買的口罩數量",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "購買成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "狀態訊息",
                                            "type": "string",
                                            "example": "購買成功"
                                        },
                                        "transaction": {
                                            "description": "交易詳情",
                                            "properties": {
                                                "transaction_id": {
                                                    "description": "交易 ID",
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "transaction_date": {
                                                    "description": "交易日期時間",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2021-01-13 01:18:23"
                                                },
                                                "transaction_amount": {
                                                    "description": "交易總金額",
                                                    "type": "number",
                                                    "format": "decimal",
                                                    "example": 12.99
                                                },
                                                "status": {
                                                    "description": "交易狀態",
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "mask_id": {
                                                    "description": "購買的口罩 ID",
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "pharmacy_id": {
                                                    "description": "銷售的藥局 ID",
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "price": {
                                                    "description": "口罩單價",
                                                    "type": "number",
                                                    "format": "decimal",
                                                    "example": 12.99
                                                },
                                                "quantity": {
                                                    "description": "購買數量",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "customer": {
                                            "description": "顧客資訊",
                                            "properties": {
                                                "customer_id": {
                                                    "description": "顧客 ID",
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "previous_cash_balance": {
                                                    "description": "交易前的現金餘額",
                                                    "type": "number",
                                                    "format": "decimal",
                                                    "example": 100
                                                },
                                                "new_cash_balance": {
                                                    "description": "交易後的現金餘額",
                                                    "type": "number",
                                                    "format": "decimal",
                                                    "example": 87.01
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "錯誤訊息",
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "customer_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "無此顧客 ID"
                                                    }
                                                },
                                                "pharmacy_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "無此藥局 ID"
                                                    }
                                                },
                                                "mask_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "無此口罩 ID"
                                                    }
                                                },
                                                "quantity": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "數量不能小於 1"
                                                    }
                                                },
                                                "pharmacy_mask": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "此藥局未提供該口罩"
                                                    }
                                                },
                                                "cash_balance": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "資金不足"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "錯誤訊息",
                                            "type": "string",
                                            "example": "購買失敗: 資料庫錯誤"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "統計特定日期區間內的口罩銷售數量和總金額",
                "description": "計算並回傳指定日期範圍內售出的口罩總數量和交易總金額。",
                "operationId": "getMaskSales",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "開始日期 (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "結束日期 (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "統計資料",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_masks": {
                                            "description": "售出的口罩總數量",
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "total_value": {
                                            "description": "日期範圍內所有交易的總金額",
                                            "type": "number",
                                            "format": "float",
                                            "example": 1234.56
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "start_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "開始日期欄位必填"
                                                    }
                                                },
                                                "end_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "結束日期欄位必填",
                                                            "結束日期必須等於或晚於開始日期"
                                                        ]
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}